:fire: IT WORKS
Added a bunch of basic middleware, moved to EJS (not really, the page is rendered using EJS, but it doesnt use any feature of EJS)

Finished the basic frontend, app now works, but it creates requests to the API like a seperate client. This does allow for cool functionality like CORS, so now someone better skilled at frontend can make a client for this!

Moved the Save button to be a seperate button to reduce number of request when adding tasks

Added a copy link button.

TODO:
- Add a way for user to use a custom id for for their todo list
- Add a delete button for individual tasks
- Add a delete button to delete the list itself
- Styling :flushed:
- Add some kind of link to go to the homepage
- (Maybe) A public list of most recent lists (Still don't have a GET ALL endpoint in API :sweat_smile:)
- Move my musings in my commit messages to README
- Watch anime after this ffs it's been a week I'm working on this app (I miss Nichijou :sob:)