:sparkles: Added Server Side validation, client error reports, PUT is now idempotent
Finally made PUT idempotent: PUT now creates a todo if id is invalid (Personally, I don't agree with this, but it is a more RESTful approach to PUT)

Fixed a bug :bug: where a todo would be deleted when attempting to move to an id that was already taken.

Added basic client side error reporting, need to make an error message show up instead. I'll do this after styling since it needs CSS.

TODO:
- :pencil: Write documentation for the API (because I enabled CORS)
- :flushed: Styling
- :sparkles: (Maybe) Pagination for the get_all endpoint and frontend
- :bulb: Comment Code
- :hammer: Look at hosting